+++ modules/ptya-map-v2/src/unit.rs
@@ -11,16 +11,16 @@ pub struct TileUnit(pub Zoom);
 
 impl Unit for TileUnit {}
 
-impl UnitCompatibility<f32, MapUnit> for TileUnit {
-    fn convert_value(&self, value: f32, unit: MapUnit) -> Option<f32> {
-        let tile_num = self.0.get_num_tiles() as f32;
+impl UnitCompatibility<f64, MapUnit> for TileUnit {
+    fn convert_value(&self, value: f64, unit: MapUnit) -> Option<f64> {
+        let tile_num = self.0.get_num_tiles() as f64;
         Some(value * tile_num)
     }
 }
 
-impl UnitCompatibility<f32, TileUnit> for MapUnit {
-	fn convert_value(&self, value: f32, unit: TileUnit) -> Option<f32> {
-		let tile_num = unit.0.get_num_tiles() as f32;
+impl UnitCompatibility<f64, TileUnit> for MapUnit {
+	fn convert_value(&self, value: f64, unit: TileUnit) -> Option<f64> {
+		let tile_num = unit.0.get_num_tiles() as f64;
 		Some(value / tile_num)
 	}
 }
